07_AdvancedConvolution/PySodium/requirements.txt
torchsummary==1.5.1